Refactor form components
