+++ client/src/App.js
@@ -2,6 +2,7 @@ import React from 'react';
 import { BrowserRouter as Router, Route, Routes } from 'react-router-dom';
 import Dashboard from './pages/Dashboard';
 import Home from './pages/Home';
+import Profile from './pages/Profile';
 
 export default function App() {
   return (
@@ -9,11 +10,11 @@ export default function App() {
       <Router>
         <Routes>
           <Route path="/dash" element={<Dashboard />} />
-          <Route path="/forgot-password" element={<Home form="reset"/>} />
-          <Route path="/update-profile" element={<Home form="update"/>} />
-          <Route path="/signup" element={<Home form="signup"/>} />
-          <Route path="/" element={<Home form="login"/>} />
-          <Route path="/login" element={<Home form="login"/>} />
+          <Route path="/forgot-password" element={<Home form="reset" />} />
+          <Route path="/profile" element={<Profile />} />
+          <Route path="/signup" element={<Home form="signup" />} />
+          <Route path="/" element={<Home form="login" />} />
+          <Route path="/login" element={<Home form="login" />} />
         </Routes>
       </Router>
     </div>
